Added new app - Touch Graph
Touch Graph plots the capacitive touch data of the MPR121 in real time.
This app uses web sockets to stream data to the browser, which
consequently required quite a few changes under the hood.
Flask-SocketIO proved to have a too slow update rate, so the socket
code was rewritten using the SockJS-Tornado module.